Crew-AI-Crash-course-main/requirements.txt
crewai
crewai_tools
load_dotenv
langchain-huggingface
embedchain 
transformers 
sentence-transformers